Added debug code to figure out CI failures

+++ test/e2e/utils.go
@@ -211,10 +211,12 @@ func printTestStackTrace() {
 }
 
 func undeployJaegerInstance(jaeger *v1.Jaeger) {
-	err := fw.Client.Delete(goctx.TODO(), jaeger)
-	require.NoError(t, err, "Error undeploying Jaeger")
-	err = e2eutil.WaitForDeletion(t, fw.Client.Client, jaeger, retryInterval, timeout)
-	require.NoError(t, err)
+	if !t.Failed() {
+		err := fw.Client.Delete(goctx.TODO(), jaeger)
+		require.NoError(t, err, "Error undeploying Jaeger")
+		err = e2eutil.WaitForDeletion(t, fw.Client.Client, jaeger, retryInterval, timeout)
+		require.NoError(t, err)
+	}
 }
 
 func getJaegerInstance(name, namespace string) *v1.Jaeger {
